factor out test diagnostic collector
